requirements.txt
# Inside tqqq_algo_trader/requirements.txt

alpaca-py
pandas
